backend/pyproject.toml
[project]
name = "ai-movie-backend"
version = "0.1.0"
description = "AI Movie Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.1",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.20",
    "cryptography==44.0.0",
    "celery[redis]==5.4.0",
    "redis==5.2.1",
    "httpx==0.28.1",
    "Pillow==11.1.0",
    "ffmpeg-python==0.2.0",
    "boto3==1.35.94",
]

[project.optional-dependencies]
test = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
